DateTimePicker: Add facilities to customize MonthView's appearance
Change-Id: I5297e451225d9d8dc2b4dc81fcbda6df1f33b892

Datetimepicker: Account for RTL locales when formulating Day labels

Doing a simple substring for the first character to generate a single
character day label doesn't work for right-to-left (RTL) and Chinese
languages.

Change-Id: Ic7d2a7880179326530830d8deee0f668dde335c8

[1/2] DateTimePicker: Add ability to set Month Title layout offset

Also add a utility function to determine Day label text's center for
an input set of coordinates.

Related to: Calendar: YearView now resolves to Day view (rather than
MonthView).

Change-Id: Iff3370540376dd62e55abe17e3abcbd4a31b924d

Datetimepicker: Fix 'Done' text clipping in DatePickerDialog

The text in the Done button gets clipped on the bottom when there
isn't ample space to work with. Removing the padding introduced by
the style; it isn't needed as min height is maintained by the button.

Change-Id: I871e8418e059dd034116ba2ceac37b97cc809a1b

Change-Id: I5297e451225d9d8dc2b4dc81fcbda6df1f33b892